+++ R/cache.R
@@ -1254,7 +1254,7 @@ single_cache_log <- function(key, cache) {
 #' # Get the error log, an object of class "error".
 #' error <- diagnose(my_target)$error # See also warnings and messages.
 #' str(error) # See what's inside the error log.
-#' error$calls # View the traceback. (See the traceback() function).
+#' error$calls # View the traceback. (See the rlang::trace_back() function).
 #' })
 #' }
 diagnose <- function(