+++ R/package.R
@@ -53,7 +53,7 @@
 #'   simplify topo_sort V vertex_attr
 #' @importFrom methods new setRefClass
 #' @importFrom parallel mclapply
-#' @importFrom rlang dots_list enquo eval_tidy expr quo_squash quos
+#' @importFrom rlang dots_list enquo eval_tidy expr quo_squash quos trace_back
 #' @importFrom storr storr_environment storr_rds
 #' @importFrom txtq txtq
 #' @importFrom utils compareVersion head menu packageVersion